move bases helpers to BasesUtil
