+++ src/test/scala/org/hammerlab/guacamole/alignment/AffineGapPenaltyAlignmentSuite.scala
@@ -1,7 +1,7 @@
 package org.hammerlab.guacamole.alignment
 
+import org.hammerlab.guacamole.util.BasesUtil._
 import org.scalatest.{FunSuite, Matchers}
-import org.hammerlab.guacamole.util.TestUtil.Implicits._
 
 class AffineGapPenaltyAlignmentSuite extends FunSuite with Matchers {
