+++ src/test/scala/org/hammerlab/guacamole/reads/MappedReadSuite.scala
@@ -2,8 +2,8 @@ package org.hammerlab.guacamole.reads
 
 import htsjdk.samtools.TextCigarCodec
 import org.hammerlab.guacamole.reference.ReferenceUtil
-import org.hammerlab.guacamole.util.{AssertBases, GuacFunSuite, TestUtil}
-import org.hammerlab.guacamole.util.TestUtil.Implicits._
+import org.hammerlab.guacamole.util.BasesUtil._
+import org.hammerlab.guacamole.util.{AssertBases, GuacFunSuite}
 
 class MappedReadSuite
   extends GuacFunSuite
