+++ src/test/scala/org/hammerlab/guacamole/reads/PairedReadSuite.scala
@@ -1,8 +1,7 @@
 package org.hammerlab.guacamole.reads
 
 import htsjdk.samtools.TextCigarCodec
-import org.hammerlab.guacamole.util.TestUtil
-import org.hammerlab.guacamole.util.TestUtil.Implicits._
+import org.hammerlab.guacamole.util.BasesUtil._
 import org.scalatest.{FunSuite, Matchers}
 
 class PairedReadSuite extends FunSuite with Matchers {
