+++ src/test/scala/org/hammerlab/guacamole/reads/UnmappedReadSerializerSuite.scala
@@ -1,7 +1,7 @@
 package org.hammerlab.guacamole.reads
 
-import org.hammerlab.guacamole.util.TestUtil.Implicits._
-import org.hammerlab.guacamole.util.{GuacFunSuite, TestUtil}
+import org.hammerlab.guacamole.util.BasesUtil._
+import org.hammerlab.guacamole.util.GuacFunSuite
 
 class UnmappedReadSerializerSuite extends GuacFunSuite {
