+++ src/test/scala/org/hammerlab/guacamole/util/BasesUtil.scala
@@ -1,0 +1,6 @@
+package org.hammerlab.guacamole.util
+
+object BasesUtil {
+  implicit def basesToString = Bases.basesToString _
+  implicit def stringToBases = Bases.stringToBases _
+}
